requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pycryptodome==3.21.0
Werkzeug==3.1.3
piexif==1.1.3
pytest==8.3.4
